Clear IMAGE_NAME_SUFFIX for all image types.
Setting this conditionally for a specific image type does not
work when combined with image conversions such as ".gz".

If you switch from sdimg to sdimg.gz in your IMAGE_FSTYPES variably,
you will end up with some dangling symlinks in the deploy directory:

    $ ls -ld --dereference tmp/deploy/images/*/core-image-full-cmdline* > /dev/null
    ls: cannot access 'tmp/deploy/images/beaglebone-yocto/core-image-full-cmdline-beaglebone-yocto-20180531205051.rootfs.sdimg': No such file or directory
    ls: cannot access 'tmp/deploy/images/beaglebone-yocto/core-image-full-cmdline-beaglebone-yocto.sdimg': No such file or directory

This change sets the variable without the _ override and also
removes the manual symlinking since the standard Yocto logic
handles it properly.

The upstream commit that added the ability to override the
IMAGE_NAME_SUFFIX variable is here and the commit log is
consistent with the above:
    http://lists.openembedded.org/pipermail/openembedded-core/2016-March/118593.html

Changelog: Title

Signed-off-by: Drew Moseley <drew.moseley@northern.tech>