tests: Use RM_WORK_EXCLUDE for image recipes.
In morty, there is an issue where the IMAGE_ROOTFS directory is cleaned up by
the RM_WORK processing.  For some of our tests, this causes a problem because
we will run a build, modify local.conf to require a rebuild of the mender artifact.
and then try another build.  In this case, the IMAGE_ROOTFS directory does not
exist so the build fails.

This is not an ideal fix since it is specific to core-image-full-cmdline and
core-image-base but it should suffice for QA purposes.

Changelog: None
Signed-off-by: Drew Moseley <drew.moseley@northern.tech>