uboot_auto_patch: Add support for most options in Kconfig files.
U-Boot is moving towards that, so our auto-patcher better follow.

As part of this we also switch to Kconfig fragment files for passing
settings. Because the settings are stored in Kconfig defconfig files,
which have no logic in them, we can no longer embed various settings
in header files. Instead, add them to a Kconfig fragment, which will
be read by the auto-patcher. It will try to add the definitions in the
right place, be it Kconfig or the source code, so that older versions
without all settings in Kconfig will also work.

The act of adding to Kconfig is also tricky, because some settings
have dependencies on other settings. So this commit adds
add_kconfig_option_with_depends.py which looks for such dependencies
in the Kconfig files and attempts to honor them.

We also clean up definitions a bit so that they are only added by the
auto-patcher when they are really needed, so for instance we avoid MMC
definitions in the UBI case.

Let's also clean up the changelog a bit, since this project has been a
long time in the making, and it's not desirable to squash the whole
thing:

Cancel-changelog: 16596c4b27c53eb1884f77bf8869d3b2494ca8fd
Cancel-changelog: e9ac1b560fe75c46c30d35a7de2dad5719f45e44

Changelog: uboot_auto_patch: Update to support U-Boot v2018.05.

Changelog: uboot_auto_patch: Add support for new Kconfig based
defines.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>