Work around function/session fixture ordering issue when skipping tests.
Since the only_with_image fixture is (and must be) a function fixture,
but ubimg_without_uboot_env is (and should be) a session fixture, the
latter cannot depend on the former, and on recent versions of pytest,
this is enforced. So the test is skipped too late. Work around it by
checking for the existence of an appropriate image instead.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>