Work around parsing failure in grub-efi by duplicating recipe.
Unfortunately there is no way to work around the bug without
duplicating the recipe, because the error is triggered inside an
anonymous Python function, which we cannot override. We can only
shadow the recipe with a different recipe.

This commit should be reverted once openembedded-core has accepted the
patch which adds the arm grubtarget.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>