Pad partitioned images up to the alignment.
This matters mostly for the emulator, which uses the file size to
determine the size of the storage device, which must be a multiple of
its device block size. However, it might be beneficial for real
storage media as well, to make sure the final sector is cleared out
when flashing the image. May increase image size slightly, but should
compress well!

Changelog: Images partitioned with GPT or MSDOS partition tables are
now padded up to the nearest alignment specified in
`MENDER_PARTITION_ALIGNMENT`. Previously the last block might be
shorter.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>