Work around unpredictable license checksum in foreign U-Boot clones.
Even though the license is expected to always be GPLv2, the exact
contents of the license file may vary depending on the exact U-Boot
fork in question. Work around it by pointing the the checksum at a
dummy file. We expect actual license problems to be caught by the
main u-boot recipe, not by u-boot-fw-utils-mender-auto-provided.

Changelog: Fix license checksum sometimes failing in
`u-boot-fw-utils-mender-auto-provided` recipe when U-Boot fork has a
slightly different README file.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>