Switch image names from containing MACHINE to containing DEVICE_TYPE.
Changelog: Switch image names from containing `MACHINE` to containing
`MENDER_DEVICE_TYPE`. So for example, if you have a build for the
raspberrypi3 machine type, with device type of "prod_rpi3", the image
will now be called `core-image-minimal-prod_rpi3.sdimg` instead of
`core-image-minimal-raspberrypi3.sdimg`.

The main reason for doing this change is that MACHINE is not enough to
separate images from each other when multiple boot loader
configurations are involved. For example vexpress-qemu can be booted
both with U-Boot and with GRUB, and these images are not compatible
with each other. To make sure that images have distinct names, this
change is necessary.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>