Switch image names from containing MACHINE to containing DEVICE_TYPE.
Changelog: None

The main reason for doing this change is that MACHINE is not enough to
separate images from each other when multiple boot loader
configurations are involved. For example vexpress-qemu can be booted
both with U-Boot and with GRUB, and these images are not compatible
with each other. To make sure that images have distinct names, this
change is necessary.

This is a backport of the commit with the same name to sumo. The
difference is that we don't enable it globally here in sumo, but only
in our own builds.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>