Fix ordering issues between clean builds and Bitbake runs.
The problem arose because we are adding new tests that use the qemu
instance in other locations than the test_update.py file, which has
been typical up to now. Now we have a situation where qemu is launched
from a clean image, then an unclean image is built as part of another
test, and the clean image is requested again by a different fixture.
But now the image is dirty!

So we fix it by returning a function instead of an object, and users
have to call it to get the image. This allows us to do the necessary
cleanings whenever it is called, instead of just the first time.

Also switch from a symlink to copying the image to the clean location,
to avoid the symlink target getting out of date.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 9b7036a48b80e9a33f7120d44e94829fc2b4c6f0)