test_uboot_compile: Fix "nativepython not found" error in recent poky.
The error happens only in the test, and is due to commit
b91562697d1d in upstream poky/sumo branch. Our test doesn't use the
same environment as the u-boot recipe, and hence its effects were not
picked up by the test properly.

To fix it, we rework bitbake_env fixture to allow any bitbake recipe.
The fixture is used to inherit the bitbake environment in the test,
and we keep this as is. However, for the test_uboot_compile test
specifically, we invoke it using the new bitbake_env_from class
instead (which the fixture also uses), which can take a parameter
which is the recipe to inherit the environment from.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>