Make sure MENDER_FEATURES check takes DISABLE into account.
Up until now, we've had these synonyms:

* MENDER_FEATURES_ENABLE == DISTRO_FEATURES_BACKFILL
* MENDER_FEATURES_DISABLE == DISTRO_FEATURES_BACKFILL_CONSIDERED

However, this is impossible if we want to base decisions on the Mender
features, which may be in the right hand variables, and use that to
put things into DISTRO_FEATURES_BACKFILL. If we did that it would
result in infinite recursion.

Therefore, we start requiring that all the features are enabled using
MENDER_FEATURES variables only, and we add a check that this is indeed
the case. Afterwards we can base decisions on the MENDER_FEATURES
variables directly, we just need to make sure we check both ENABLE and
DISABLE.

Changelog: none

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>