Fix basehash mismatch error message on recent master.
Switch to using IMAGE_VERSION_SUFFIX instead of DATETIME, which is
what poky has also done in a65c1216d7c4c. As part of that they also
removed the `IMAGE_NAME[vardepsexclude] += "DATETIME"` that existed
before, which is why the error message started appearing. It was
reintroduced for the IMAGE_VERSION_SUFFIX variable, so that one is
safe to use.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>