tests: Add mender-image-tests as a submodule.
This is attempting to deal with the fact that image tests now live in
a separate repository. The existing method where we copy over the
tests from that repository is very fragile, and doesn't allow
specifying alternate branches, which will be important as newer
versions of mender-image-tests drift out of sync with older branches
of meta-mender. We will then need to branch mender-image-tests and
switch the submodule over to that.

One slight complication with our use of this submodule is that we want
to track branches, not explicit commits. We solve this by having an
automatic pull before testing starts, which can be disabled if you're
debugging something locally, or in the Jenkins case where we will
check out the exact commit that was specified in the build.

The "repo" tool was considered as an alternative to submodules, but I
didn't see enough benefit, given that repo is very workflow oriented,
and this submodule is not really meant to be used by users very much
(hence the "ignore" attribute in the .gitmodules file). This also has
the benefit of not requiring another tool.

Changelog: None
Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>