meta-mender-toradex-nxp: fix fw_env.config override
We provide a custom fw_env.config file and we are trying
to override the one provided by u-boot-toradex-fsl-fw-utils.

The upstream location is:

    meta-toradex-nxp/recipes-bsp/u-boot/files/colibri-imx7/fw_env.conf

and in meta-mender we have:

    meta-mender-toradex-nxp/recipes-bsp/u-boot/files/mx7/fw_env.config

This does not result in an override and the upstream fw_env.config
file is used. The full path must match for it to be considered an override.

Changelog: Fixed wrong fw_env.config used for Toradex iMX7 SoM.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>