tests: Work around corrupted UBI volume on partial write.
Corrupting the volume when the write doesn't write everything it's
supposed to, is an intentional security feature of UBI, and it can
trigger when we do checksum mismatch tests, which suddenly stop
writing when they discover that the checksum is wrong. This apparently
is checked more rigorously in recent kernels, so accept a non-readable
volume when performing UBI tests.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>