Fix images not being modifiable by mender-artifact in thud and later.
Changelog: Title

The problem is that the "metadata_csum" feature is very recent, and in
fact even the version shipped in poky (1.44.3) doesn't support it
(both fsck.ext4 and tune2fs refuse to handle it). Why they turned it
on in the first place is a mystery to me. In any case it should be
safe to turn off for our images, so that debugfs (and hence
mender-artifact) is able to handle the image.

If anyone requires this feature, it can be flipped on again on the
built image by a recent version of e2fsprogs/tune2fs.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>