part-images: add missing u-boot:deploy dependency for ARM
The current logic only adds a dependency to u-boot:deploy when
mender-uboot feature is enabled. This becomes problematic when
GRUB integration is used on ARM (which is now default), because
here we actually depend on both U-boot and GRUB.

You typically hit issues with missing dep when MENDER_IMAGE_BOOTLOADER_FILE
is set and the image generator can not find the necessary files.

Changelog: Title

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>