grub: ensure "test" module is builtin
Mender generates a grub.cfg that utilizes `[` which is an alias for `test` and
we should probably make sure that it is builtin when grub-mender.inc is used.

This has probably gone unnoticed because the upstream recipe includes this by
default.

I experienced problems when working with a downstream grub-efi recipe,
which did not include the `test` module in GRUB_BUILDIN.

Changelog: Title

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>
(cherry picked from commit d260495f76aedc59166fc5ad1bdbe3eecc6d25ba)