Fix failure when submitting a PR with a branch in the main repo.
The problem is that the version detection mistakes the custom branch
for a Yocto branch, so be explicit that only the listed branches are
candidates for Yocto branches. This means we don't have to handle the
master branch specially anymore.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 4e7d8f6af6a2911c9423ffa43259bb00b80df568)