mender: Setup Live installer with HDDIMG
The current HDDIMG (and ISO for that matter) support in Yocto sets up multiple
bootloader options. One is standard live boot but the other is an install mechanism
so you can install your build onto a target without removable media. This patch
provides a version of this scripting that works with Mender. Basically, the SDIMG
UEFIIMG or MTDIMG file are copied into the HDDIMG and the initrd runs a script
which uses 'dd' to write the image.

This has been tested on an X86 platform with an internal MSATA drive and an external
USB drive.

Changelog: Title

Signed-off-by: Drew Moseley <drew.moseley@northern.tech>
(cherry picked from commit fa78acf5a0a6d025fe9c2f2f5f8c144454c2a8e0)