raspberrypi: split RPi3 and RPi4 integration
meta-raspberrypi carries U-Boot patches to support RPi4, and these
patches are only applied when building 'raspberrypi4' machine. For this
reason we can not have one integration patch for both RPi3 and RPi4 as it
will not apply when building for a RPi3 target.

We have to split out the RPi4 support in a individual patch that is only
applied when buildingfor RPi4.

This will sync-up in future releases once RPi4 support is part of U-Boot
upstream code.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>