mender: Use += rather than _append for IMAGE_FSTYPES.
Using _append breaks initramfs builds and attempts to build Mender style
images for them.  Basically, the initramfs image files set a new value
of IMAGE_FSTYPES but when we use _append we are applied _after_ that;
using += makes sure that the override done in the image recipe takes
place.

More details can be seen in this discussion:
    https://github.com/ndechesne/meta-qcom/issues/61

Changelog: Title
Signed-off-by: Drew Moseley <drew.moseley@northern.tech>