Fix incorrect partition types after we started using `rawcopy`.
This was introduced in 30a27a9f5e9ba8edb1e879380a64fc8cc994af43. When
we switched away from the `rootfs` source, we lost the identity of the
filesystem, and all partitions became the generic "Microsoft basic
data" type. We fix GPT by using `--part-type` argument, but the MBR
code in wic doesn't support using it, so we need to fix those entries
manually with fdisk.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>