Ensure that preapre/cleanup_excluded_directories are added to ARTIFACTIMG_FSTYPE
The current logic only parses the content of IMAGE_FSTYPES and adds the tasks
to the images listed there.

But ARTIFACTIMG_FSTYPE could actually be brought in trough IMAGE_TYPEDEP_mender or
e.g IMAGE_TYPEDEP_sdimg and in this case the excluded directories are never cleaned
up.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>