Ensure that preapre/cleanup_excluded_directories are not added to conversion images
A simple example would be if you had the following:

    IMAGE_FSTYPES = "tar.gz"

The current logic would add preapre/cleanup_excluded_directories functions to
the "do_image_tar.gz" which is not a valid task and the actual task "do_image_tar"
would not have had the excluded directories cleaned-up.

For this reason, we strip away suffixes from images names in IMAGE_FSTYPES so
that the correct task gets the exclude functions.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>