Add patch to grub-editenv-mender to workaround configure errors
For some uknown reason, in Yocto warrior the grub configure script is
not able to find two header files: wchar.h and sysmacros.h, which
triggers the build to abort with compile errors.

The header files are actually there for glibc and probably most of libc
versions used for native compile. As we are interested in grub-editenv,
and not compiling grub itself, it should be safe to assume tha the
headers are present in the sysroot for the toolchain.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>