Apply openssh openssl10 dependency to all configurations
A considerable amount of tests are failing due to tiemout related
issues, and even some jobs are timing out after 8 hours of build time.
The exception is on vexpress-qemu-flash config, where all tests run with
no problem.

Hence, we suspect that from warrior on, the random numbers generation
issue in openssl is affecting all our QEMU configurations, so let's
stick to openssl10 for all of them.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>