Fix and strict check for mender-migrate-configuration MEN-2757
In the one hand, fix the recipe itself by moving the cp statement to
do_compile as required by mender-state-scripts class.

In the other hand, this recipe will only work if split-mender-config is
disabled in the same build. Add a hard check for this making it depend
on mender and document the improve the usage doc in the changelog entry.

Cancel-changelog: 5acbed272640d72fcd6fd6d6043c434b2f0fc2a1

Changelog: Splits the mender.conf configuration file into a transient
configuration /etc/mender/mender.conf and a persistent congiguration in
/data/mender/mender.conf. This split is enabled by default, and it can
be opt-out by adding `PACKAGECONFIG_remove = "split-mender-config"` to
local.conf.

Changelog: Add a new recipe mender-migrate-configuration to ease the
migration of devices with single mender.conf to the new setup of split
configuration files. To build an updgrade for such device, add the
recipe with `IMAGE_INSTALL_append = " mender-migrate-configuration"`,
disable the split feature (for this update only) with
`PACKAGECONFIG_remove = " split-mender-config"`, and specify which
parameters to migrate (at least the partition parameters) with
`MENDER_PERSISTENT_CONFIGURATION_VARS = "RootfsPartA RootfsPartB"`.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>