test_uboot_automation: Update failed board count for Flash tests.
The whole thing is an inexact science, but it looks like the majority
of failing boards are because a few select boards have been added with
different configurations, so they keep failing over and over. I could
not spot any serious regressions, so updating the test number instead.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>