Add handling of OSError exception while connecting to QEMU.
It seems that while the QEMU is not yet up and running
`paramiko.client.SSHClient.connect()` method is returning OSError
`Cannot assign requested address` while the other end of the connect
socket is not yet initialized. We need to handle it gracefully and
retry after QEMU is up and running.

Changelog: none

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>