Update min_mender_version sentinel to 2.2.0 for artifact deps tests
The current mender-version guard had a version of 2.0.0, which is too low, since
the changes in mender-artifact which holds the depends and provides
functionality is only available in master at the moment of this writing. Hence
set the tag to 2.2.0, which for now only includes the master branch.

Without this fix, cherry-picks to older branches will fail the acceptance tests,
if they employ older versions of mender-artifact.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>