Add basic entry point for upgrading tests to Python3.
As we are doing the upgrade to Python3 we also need to upgrade Fabric
to the Fabric2.x version and pytest to pytest > pytest4.x.

This comes with many changes:
* different way of handling the connections
* lack of global config object for handling CLI parameters
* some Python2 -> Python3 upgrade logic changes

As the first step we are adding new Connection object for handling
Fabric connections to the tested host. We are also adding some needed
custom test marks and Python3 dependencies.

Changelog: none

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>