Handle pytest change of pytest.config global object deprecation.
As of pytest4.x pytest.config is deprecated and in pytest5.x it
is deleted completely. In order to get the CLI parameters passed over
to the tests we need to use fixtures.
As handling conversion tests CLI parameter would cause a lot of
refactoring we are using hack to maintain the global config object.

NOTE: @pytest.mark.conversion was added to the image tests as the
counterpart of `pytest_collection_modifyitems` function.

Changelog: none

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>