Add requirements_py3.txt file for Python3 requirements support.
As we need a way to tell if we want to run Python3 or Python2 tests
we are using a `requirements_py3.txt` file as a switch. If the file
is detected by the test framework all the tests will be executed
using Python3. Otherwise Python2 will be used.

Changelog: none

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>