mender-image: Add DEPENDS to include WKS_FILE_DEPENDS.
This is used to add extra dependencies needed for wic images on certain platforms.
We are using wic but not explicitly so the upstream detection of this does not
work.  Notably on certain Toradex platforms this is needed for some of the
U-Boot logic.

Changelog: Title
Signed-off-by: Drew Moseley <drew.moseley@northern.tech>
(cherry picked from commit 49c361ea7f688735c14a5821a07ce230b334fb98)