Work around missing LSB support in zeus.
This is a very intrusive change, which creates the `/lib64` symlink on
the root filesystem for 64-bit systems. This symlink is normally
missing when building with Yocto, but was provided by the LSB package
previously. The directory is necessary to remain compatible with
software built outside of Yocto. The Mender demo artifact, as well as
the binary delta update module are examples of such components.

If the link turns out to conflict with another package, it should be
possible to work around the problem by adding this to the other
package, either in the original `.bb` file or a `.bbappend` file:

```
RPROVIDES_${PN} += "lsb-ld"
```

Changelog: Commit

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>