mender: Ensure excluded directories work for all image types.
This commit changes the functionality of the directory exclusion to simply backup
the existing one and restoring it when done.  The previous version relied on modifying
the IMAGE_ROOTFS variable but that did not work for image commands such as IMAGE_CMD_tar
since that is defined as a variable rather than a function and would be fully expanded
before we modified the IMAGE_ROOTFS variable.

Signed-off-by: Drew Moseley <drew.moseley@northern.tech>

Changelog: Title