tests: Switch `setup_board` to session fixture.
This is how it used to be in the Python 2 era, and greatly increases
the testing speed by not having to relaunch the client for every test.
It does mean that each test has to clean up properly though, but this
was also the case before Python 3 was introduced, so it should be a
safe assumption.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>