raspberrypi: add state script to update boot firmware
Raspberry Pi boards have a set of boot firmware files that are
located on the vfat boot part, and these files are not updated
when you perform an update of the root filesystem.

Occasionally there will be changes to the Raspberry Pi software stack
that requires that these files are update. Typically there is something
in the Linux kernel that requires something that is in the boot
firmware. This means that to update the Linux kernel you must also
update the boot firmware files.

Above is really sub-optimal design of the Raspberry Pi boards but a
limitation that we must live with.

Note that the DTB files are also part of "boot firmware" and included in
the state script update.

The provided state-script can be enabled by adding the following to e.g
local.conf:

    INHERIT += "rpi-update-firmware"

The ArtifactInstall_Leave_50 script can be overriden to customize what
files to update, e.g only DTB files. By default all files on the boot
part will be updated including config.txt and cmdline.txt.

NOTE! Updating the boot firmware files can not be done atomically and the
files are not roll-backed even though an rootfs rollback is performed. Conclusion,
this is an risky operating which could brick your device.

Changelog: Commit

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>