Workaround for warrior: Rename mender-client.service to mender.service.
We don't want this potential breaking change on warrior and older.
Doing the change with a conditional ensures that it will work
correctly both for old and new versions.

Also provide a patch for clients version 2.3 and later that makes sure
to send a signal to both service names, since the service name on
warrior won't match what the client expects.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>