Added a Mender client append file to enable coverage analysis
This commit adds a mender_client%.bbappend file which will automatically
instrument the source code of Mender clients newer than 2.2.x with coverage
functionality.

This is done through expanding the 'do_configure' step in the original recipe
with a few points of note:

1) The function is not idempotent, so that a configuration must always be done
on a clean source repository, therefore a 'configure_prepend()' is added in
which the source tree is deleted if it is not a symbolic link (this we cannot
recursively delete, as it may point to 'externalsrc'.

2) It inverts (i.e., removes) the symbolic link created in the 'do_configure'
step already present, as inherited from 'go.bbclass'.

3) It uses 'rsync' to copy the Mender source files from the source folder ${S}
to the build folder ${B}.

Note that these features will never be enabled for older clients. This is due to
the 'instrument-binary' target which needs to be present in the Makefile of the
Client.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>