MEN-3477: Start downloading grub-efi from grub-mender-grubenv URLs.
The URLs include the revision SHA of grub-mender-grubenv as well as
the version of GRUB, which means we can get a stable checksum instead
of a constantly changing one.

Also standardize on some things that are shared between the
grub-efi-mender-precompiled and grub-mender-grubenv recipes, such as
URIs, revisions and the GRUB module list. We also check that the GRUB
module list defined in Yocto is the same as the one used to compile
the binaries.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>