mender-client: make sure that mender-grow-data.service only runs once
To be able to achive the desired functionality, logic needed to be
moved to a seperate script, even though I would have liked to avoid
it.

This change was primarly motivated by the addition of the "fdisk w"
call, which ensures that GPT backup headers are written. But this
call has other implications since this service runs on each boot:

1. On each boot it will run "fdisk w", which is an partition table
   alterning command. Even if the underlaying layers might support
   this, it feels unecessary.

2. Altering partition table on GPT is safe, because of the backup
   headers but this service will also run on MBR style disk images,
   where there is no backup

Changelog: Make sure that mender-grow-data.service only runs once.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>