mender: do not rely on x-systemd.growfs in /etc/fstab to resize fs
systemd will generate a service file if it finds the
"x-systemd.growfs" flag in /etc/fstab on a specific part. The
problem with the generated service file is that it contains the
following depedency:

    Before=local-fs.target

local-fs.target is an indication that the rootfs is mounted and ready
, essentially all services are dependent on this one. This means
that boot is blocked, until the resize of the data part has
completed and this can take several minutes.

The dependency on local-fs.target would make sense if we where
resizing the rootfs filesystem image, but we are not. This has
been reported upstream and there are plans to resolve this, but
until then we will provide a customer service file which is a copy
of the generated one but with the local-fs.target dependecy removed.

Changelog: Boot no longer blocks while the data part is resized, and
instead this is a background process.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>