Enable installation to eMMC and dual booting
To add a live installer to the image, this requires two steps:

    * Append "mender-emmc-install" to MENDER_FEATURES_ENABLE
    * Set MENDER_INSTALL_DEVICE to "/dev/mmcblk1"

U-Boot will try to boot from the SD card if present, and from eMMC if
there is no SD card.

TODO:

Several device paths in the image are hard-coded to /dev/mmcklb0. On first
boot, they are updated to /dev/mmcblk1. The files are:

    * /etc/fstab (requires hook prior to loading systemd)
    * /data/mender/mender.conf
    * /data/u-boot/fw_env.config (if using U-Boot for A/B partitioning)
    * /boot/efi/EFI/BOOT/grub.cfg (if using GRUB; updated at install time)

Changelog: Title

Signed-off-by: Garrett Brown <garrett.brown@aclima.io>