MEN-3513: Work around unstable checksum when using 64bit ext4 feature.
See the ticket for details. The underlying checksum on disk doesn't
actually change, but it changes in memory, and it doesn't seem to be
possible to read the correct checksum out of the block device. Since
we need this for mender-binary-delta to work, just disable the 64bit
feature.

Changelog: Disable `64bit` ext4 filesystem feature.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>