uboot_auto_configure: build U-Boot the same way Yocto does
The uboot_auto_configure.sh script uses the cross compiler to build
cmd/version.o from the U-Boot sources so it can generate a list of files to
patch. However, this will fail if the target board (and therefore the target
toolchain) uses floating point flags (e.g. armv5tehf-vfp) since U-Boot is not
meant to be built with floating-point.

The Yocto recipes invoke the cross toolchain when building U-Boot in a way
so that any floating-point build flags are not passed to the build. Tweak
the way the cross toolchain is invoked by following the way Yocto invokes
it.

Changelog: Title

Signed-off-by: Trevor Woerner <twoerner@gmail.com>