test_uboot_compile: Execute defconfig step to identify ARM boards.
Previously we searched through the files to try to detect ARM boards,
but this is error prone, and especially ARM64 boards are difficult to
detect. These are becoming more common, so we need another way.

So instead we execute the defconfig step of every board, and then look
for `CONFIG_ARM=y` inside the resulting configuration. This takes a
lot longer than the file searching, but is much more accurate. Given
that it eliminates more than a hundred false positives, it is well
worth it since we no longer have to compile those, and the total
testing time still should go down.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>