tests: Fix U-Boot compile test not running for pull requests.
We need to adapt the conditions for Gitlab, which uses "pr_*" branches
instead of "pull/*" branches. The conditions need to figure out that
this is a pull request, and it should therefore run the tests. This
has been broken for a while, but due to the timestamp condition, in
most cases it still works. However for PRs that are tested more than a
week after their commit date, the test may be erronously skipped.

In addition, the path to the U-Boot recipe directory has also been
wrong for a long time, causing only test fixes to be picked up, not
recipe fixes.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 51b58c824eb4f7a59ead0eeed4f0f3d519af98bf)