uboot_auto_configure: Handle unwanted config options better.
Actively disable CONFIG options we want to get rid of instead of just
removing them from the defconfig file. This helps in the case where we
try to remove an option, but it's immediately brought back because the
default is to be on.

Changelog: U-Boot auto-configuration: Better algorithm for removing
options from defconfig files. This increases board compatibility.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>