Fix broken `set-time-epoch` feature in dunfell.
On dunfell it is apparently not enough anymore to just specify
`set-time-epoch` to force systemd to use 1st Jan 1970 as epoch. It is
apparently possible to to use any epoch you like, and the default is
the build time, so we need to specifically set it to zero. See the
comment above the change for context on why the epoch is needed in the
first place.

Changelog: Fix fsck running on every boot in dunfell.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>