Fix PACKAGECONFIG not propagating RDEPENDS properly.
We have a conditional dependency on bash when the "modules"
PACKAGECONFIG is enabled. This is added to the RDEPENDS variable by
some complex bitbake magic via the PACKAGECONFIG[modules] string we
have set. It seems like if PACKAGECONFIG is appended to conditionally
using "_append_mender-client-install", then the RDEPENDS population
stage doesn't work.

Fix by evaluating the feature toggle inline instead of via a selector.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>