MEN-3912: Disable filesystem journal on read-only-rootfs.
There is evidence from users of mender-convert that having the
`has_journal` feature turned on in the filesystem features causes fsck
to break the checksum of the filesystem, even when it doesn't make any
changes. Currently the Yocto-produced image does not run fsck, so it
is not a problem there, but it could be if someone uses an initrd for
example (this is where fsck is run in the mender-convert case). So
this fix is simply about being proactive and disabling it now, before
it becomes a problem.

Changelog: Disable filesystem journal on read-only-rootfs, which
sometimes causes unstable rootfs checksum together with fsck.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit b3b4464663effa58abf51eb08df5b5ca09c9f04e)