Make mender-features and OVERRIDES handling always run.
OVERRIDES handling needs to be in a class which is always included, or
it won't work properly. This should have no effect unless mender
features are actually enabled, since it filters by this keyword. In
this case the class which contains the function would have been
included anyway, so the result is the same, we are just making sure
the evaluation order is correct.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>