Remove e2fsprogs patch to fix bmaptool issue.
This patch, which removes  metadata_csum from the available ext4 features, breaks flashing with bmaptool.
It was created at a time when tools such as fsck.ext4 and tune2fs did not handle that feature.
Nowadays, these tools have been updated and it is safe to remove the patch.

Changelog: Title

Signed-off-by: Pascal Hache <hacpa@touchtunes.com>