Fail build upon missing demo WiFi password and/or SSID
Do not allow empty WiFi password and/or SSID to silently be set in
the wpa_supplicant configuration to an empty string when the demo layer is used.

Changelog: Do not allow empty WiFi password and/or SSID to silently be set in the wpa_supplicant configuration to an empty string when the demo layer is used.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>